caching workflow and update all ecutwfc and ecutrho to expect int

+++ aiida_sssp_workflow/workflows/convergence/delta.py
@@ -93,8 +93,8 @@ class ConvergenceDeltaWorkChain(BaseLegacyWorkChain):
             "element": orm.Str(self.ctx.element),  # _base wf hold attribute `element`
             "configuration": orm.Str(self.ctx.configuration),
             "pw_base_parameters": orm.Dict(dict=self.ctx.pw_parameters),
-            "ecutwfc": orm.Float(ecutwfc),
-            "ecutrho": orm.Float(ecutrho),
+            "ecutwfc": orm.Int(ecutwfc),
+            "ecutrho": orm.Int(ecutrho),
             "kpoints_distance": orm.Float(self.ctx.kpoints_distance),
             "scale_count": orm.Int(self.ctx.scale_count),
             "scale_increment": orm.Float(self.ctx.scale_increment),
